More testing with menu features

+++ src/Interfaces/BuyingPlants.java
@@ -3,5 +3,5 @@ package Interfaces;
 import com.company.Plants;
 
 public interface BuyingPlants {
-    void buyPlant(Plants plants, Double amount);
+    void buyPlant(Plants plants, int amount);
 }
